# contracts/camera-stream.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus Camera Stream API
  version: 1.0.0
  description: Contract-first API for team-camera. Camera Stream receives frames, exposes frame history, calls AI Vision for detection, and publishes camera events to Analytics.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:4010
    description: Prism mock server
  - url: http://localhost:8000
    description: Local Camera Stream service
tags:
  - name: health
    description: Service readiness and dependency configuration
  - name: frames
    description: Camera frame ingestion, lookup, and analysis trigger
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: getCameraHealth
      tags: [health]
      summary: Check Camera Stream health
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
              examples:
                ok:
                  value:
                    status: ok
                    service: camera-stream
                    version: 1.0.0
                    dependencies:
                      vision: http://team-vision:8000
                      analytics: http://team-analytics:8000
        "500":
          $ref: "#/components/responses/DependencyError"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/frames:
    post:
      operationId: uploadFrame
      tags: [frames]
      summary: Upload a camera frame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrameCreate"
            examples:
              motion:
                value:
                  camera_id: CAM-A01
                  location: Main lobby
                  frame_format: jpeg
                  image_base64: dGVzdC1pbWFnZS1mcmFtZQ==
                  captured_at: "2026-05-13T08:30:00+07:00"
                  motion_score: 0.82
      responses:
        "201":
          description: Frame accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrameAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      operationId: listFrames
      tags: [frames]
      summary: List accepted camera frames
      parameters:
        - name: camera_id
          in: query
          required: false
          schema:
            type: [string, "null"]
            minLength: 3
            maxLength: 40
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Frame list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FramePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/frames/{frame_id}:
    get:
      operationId: getFrame
      tags: [frames]
      summary: Get one frame by id
      parameters:
        - $ref: "#/components/parameters/FrameId"
      responses:
        "200":
          description: Frame detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/frames/{frame_id}/analyze:
    post:
      operationId: analyzeFrame
      tags: [frames]
      summary: Send frame to AI Vision and publish Analytics event
      parameters:
        - $ref: "#/components/parameters/FrameId"
      responses:
        "200":
          description: Frame was analyzed through dependencies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/DependencyError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: lab-token
  parameters:
    FrameId:
      name: frame_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^FR-[0-9]{8}-[0-9]{4}$"
  schemas:
    HealthStatus:
      type: object
      additionalProperties: false
      required: [status, service, version, dependencies]
      properties:
        status:
          type: string
          enum: [ok]
        service:
          type: string
          example: camera-stream
        version:
          type: string
        dependencies:
          type: object
          additionalProperties:
            type: string
    FrameCreate:
      type: object
      additionalProperties: false
      required: [camera_id, location, frame_format, image_base64, captured_at, motion_score]
      properties:
        camera_id:
          type: string
          pattern: "^CAM-[A-Z][0-9]{2}$"
        location:
          type: string
          minLength: 2
          maxLength: 120
        frame_format:
          type: string
          enum: [jpeg, png]
        image_base64:
          type: string
          minLength: 16
          maxLength: 500000
        captured_at:
          type: string
          format: date-time
        motion_score:
          type: number
          minimum: 0
          maximum: 1
    FrameAccepted:
      type: object
      additionalProperties: false
      required: [frame_id, camera_id, accepted, motion_level, created_at]
      properties:
        frame_id:
          type: string
        camera_id:
          type: string
        accepted:
          type: boolean
        motion_level:
          $ref: "#/components/schemas/MotionLevel"
        created_at:
          type: string
          format: date-time
    MotionLevel:
      type: string
      enum: [none, low, medium, high]
    CameraEvent:
      oneOf:
        - $ref: "#/components/schemas/FrameCapturedEvent"
        - $ref: "#/components/schemas/FrameAnalyzedEvent"
      discriminator:
        propertyName: event_type
        mapping:
          camera.frame_captured: "#/components/schemas/FrameCapturedEvent"
          camera.frame_analyzed: "#/components/schemas/FrameAnalyzedEvent"
    FrameCapturedEvent:
      type: object
      additionalProperties: true
      required: [event_type, frame_id, camera_id, captured_at]
      properties:
        event_type:
          type: string
          const: camera.frame_captured
        frame_id:
          type: string
        camera_id:
          type: string
        captured_at:
          type: string
          format: date-time
    FrameAnalyzedEvent:
      type: object
      additionalProperties: true
      required: [event_type, frame_id, camera_id, occurred_at]
      properties:
        event_type:
          type: string
          const: camera.frame_analyzed
        frame_id:
          type: string
        camera_id:
          type: string
        occurred_at:
          type: string
          format: date-time
    FramePage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
    AnalyzeResult:
      type: object
      additionalProperties: true
      required: [frame_id, vision, analytics]
      properties:
        frame_id:
          type: string
        vision:
          type: object
        analytics:
          type: object
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: [string, "null"]
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationError:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Frame not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    DependencyError:
      description: Dependency returned an error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    DependencyUnavailable:
      description: Dependency timeout or unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"